+++ src/User_Info.java
@@ -1,7 +1,7 @@
 public class User_Info {
     public static String gender;
     public static Float weight;
-    public static int height;
+    public static Float height;
     public static Float weight_goal;
     public static String username;
     public static String password;
